requirements.txt
pymongo==4.8.0
python-dotenv==1.0.1
streamlit==1.37.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
nltk==3.9.1
textblob==0.18.0.post0
faiss-cpu==1.8.0
transformers==4.42.4
torch>=2.2.0
sentencepiece>=0.2.0
tqdm>=4.66.4
langchain==0.2.11
langchain-community==0.2.10
openai==1.42.0
google-generativeai==0.7.2
pydantic==2.8.2
matplotlib==3.9.0
streamlit-webrtc==0.47.1
uvicorn==0.30.3
fastapi==0.111.1
